+++ modules/euclidean_distance.cpp
@@ -5,7 +5,7 @@
 #include <map>
 using namespace std;
 
-void euclidean_distance(vector<vector<float>>&vec,vector<float>query_point,map  <int,float>&distances) { 
+void euclidean_distance(vector<vector<float>>&vec,vector<float>query_point,map  <int,double>&distances) { 
     double euclidean=0; 
     for(int i=0;i<int(vec.size());i++){ //getting the number of total vectors
         for(int j=0; j<int(vec[i].size());j++){
